Expose Client request encoding from an endpoint
The main difference between MockNetworkURLProtocol v2 and v3 is that in
 the `canInit(with:)` v2 method compares endpoint paths and v3 compares
the whole URL so now in order to use MockNetworkURLProtocol it’s
necessary to know the whole URL for each endpoint. In v3 that logic was
refactored in a `RequestEncoder` protocol that is accessible from the
`APIClient` class but in v2 that logic is private in the `Client`.

This commit exposes the `Client.encodeRequest(for:)` so it can be used
in unit tests to mock URLRequest's.