Compare only significant parts of the URL
The new MockNetworkURLProtocol implementation compares URL's in order
to know if a request should be mocked or not. The problem with this
approach is that two semantically identical URL's can be syntactically
different like in https//a.b.c/d?e=f&g=h and https//a.b.c/d?g=h&e=f

This commit tries to address this issue by only comparing significant
parts of an URL that should stay the same but also sufficiently
identify the URLRequest we are trying to mock.