Fix race condition in `ChannelListController_Tests`
The original test implementation wasn't correct. The `write` call caused the change aggregator to be called + we called it manually again. When the system was under heavy load, the changes to the main thread weren't properly propagated by the time we called the changes callback.

This implementation doesn't call the callback manually and relies only on the callback from the change aggregator. Technically, this is less atomic because we rely on the change aggregator to work correctly for the test to pass. However, in the case it makes sense. We would have to mock the aggregator to do this properly which seems like too much work atm.