[CIS-280] Backport URLProtocol mock and recorder classes
* Replace v2 URL mock helpers by v3 versions

Copy and paste the MockNetworkURLProtocol and
RequestRecorderURLProtocol v2 implementations by the ones found in v3
as v2 has a less convenient version of this API and it lacks some
functionality.

* Fix Atomic syntax in v2 MockNetworkURLProtocol

The v3 MockNetworkURLProtocol uses an Atomic property wrapper
implementation that is not available in v2, but in v2 we have an Atomic
implementation based on a class so this commit is adjusting from the
v3 Atomic to the v2 implementation.

* Expose Client request encoding from an endpoint

The main difference between MockNetworkURLProtocol v2 and v3 is that in
 the `canInit(with:)` v2 method compares endpoint paths and v3 compares
the whole URL so now in order to use MockNetworkURLProtocol it’s
necessary to know the whole URL for each endpoint. In v3 that logic was
refactored in a `RequestEncoder` protocol that is accessible from the
`APIClient` class but in v2 that logic is private in the `Client`.

This commit exposes the `Client.encodeRequest(for:)` so it can be used
in unit tests to mock URLRequest's.

* Init Client Device tests URL mocking as in v3

MockNetworkURLProtocol v2 implementation has been replaced with v3
so this commit applies the same initialization in tests as it's done
in v3

* Fix Client Device unit tests build errors

The v2 MockNetworkURLProtocol implementation only needed to know
the endpoint to mock URLRequests but as it has been replaced with v3
now we need to encode ourselves the URLRequest using the Client.

* Compare only significant parts of the URL

The new MockNetworkURLProtocol implementation compares URL's in order
to know if a request should be mocked or not. The problem with this
approach is that two semantically identical URL's can be syntactically
different like in https//a.b.c/d?e=f&g=h and https//a.b.c/d?g=h&e=f

This commit tries to address this issue by only comparing significant
parts of an URL that should stay the same but also sufficiently
identify the URLRequest we are trying to mock.

* Fix formatting

* Add helper to compare two JSON serializations

* Wait for a matching network request

Instead of checking only the first network request,
wait for a matching request in the given timeout period

* Renable previously randomly failing tests

* Rename assert network request as assert extension

* Make Assert.networkRequest callable from AssertAsync

* Reuse AssertJSONEqual code

* Print list of recorded URLRequest's

* Display also requests matching by path

* Add shorthand API for network request assertion

* Improve comments

* Improve documentation

* Improve error message formatting

* Fix typo