Fixes ChannelController.startUpdating completion called multiple times
As it is, ChannelController.startUpdating passes its completion param to multiple places:
- to setupDatabaseObservers
- to channelCreatedCallback
so, if an error occurs in multiple places, it'll be reported in multiple places. If this is what we want, then it's correct.
Another issue is that `startDatabaseObservers` should not be called before channel is actually created (so `isChannelAlreadyCreated` should be true) but it's not checked. I've also refactored it so `startDatabaseObservers` and `setupEventObservers` are called once, in the correct place: if channel exists, called in `startUpdating` otherwise called in `channelCreatedCallback` (hence in set(channelQuery:))