Update `deleteMessage` logic in `MessageUpdater`
We get the crash trying to map `MessageDTO -> ChatMessage` when we catch
delete event by FRC because after the deletion the relationships
become nil and we cannot parse `author: ChatUser`.

Soft delete is nice workaround. For local messages the same behavior as
for sent messages is applicable: the message stays in the UI but with
`The message is deleted` content and `is only visible to you` indicator