Allow message edit/delete from all users
Message edit/delete is a user permission so some users are able to edit/delete other's messages. Backend controls this, so we should let backend return an error. By default, admin/moderators have this permission, but this configurable and it's not sent by backend, so safest to let users control this logic on their end.