Fix Combine issues when creating .swiftinterface
When creating an xcframework out of StreamChatUI, Swift fails to properly generate the .swiftinterface for it.
In places where we are using SwiftUI, it complains about missing Combine imports.
This is due to the fact that SwiftUI uses Combine, so the declaration of some of its types include references to
Combine. Even though it seems like a bug, this should fix it for now.

Extract from the generated .swiftinterface:
`StreamChatUI.ChatChannelListItemView, Combine.ObservableObject where Content :`