Let RequestEncoder fakely succeed to catch connection errors
The most probable reason why a RequestEncoder can timeout when waiting for token or connectionId is
because there's no connection. When returning an error, it will just fail without giving an opportunity
to know if the timeout occured because of a networking problem.
Instead, now, returning a success, even without the needed token/connectionId, it will account for this
case and instead will let the APIClient return a connection error if needed.
That way, the request can be requeued, and we give a much better experience.