Fix decoding failure for channel.hidden
This only happens for channel.hidden events coming from sync endpoint. Backend explained that `clearHistory` is treated as custom data, and custom data is not sent on `/sync` call, so it's not sent. It's a bug that will be resolved by backend, but client needs a workaround meanwhile.

Also resolves that decoding errors are not printed to console.