Fix unwatched channels are not hidden on new message
When the client starts for the first time, the channels linked to the query will be watched. If a channel was hidden on a previous session, that channel will not be linked to the query, so will not be watched. We won't receive a `MessageNew` event for that channel (since it's not being watched) and we won't mark it as `hidden = false`, so it won't show up.

The visibility middleware should also act on `Notification` events and mark the channel accordingly. Then it can be linked to any query that's currently alive.