Multiple fixes for channel hiding/appearing logic (#2056)
* Fix DemoApp keeping ChannelVC alive

We were always using splitVC, which keeps the ref to latest displayed VC, which keeps it alive. We should use the old logic and only use splitVC when we detect an iPad

Keeping a ref to ChannelVC causes the latest displayed channel to keep marking newly received messages as "read", which messes up the unread count

* Add hiding channel with clear history option to DemoApp

* Remove repeated pinned message saving in MessageDTO

* Fix decoding failure for channel.hidden

This only happens for channel.hidden events coming from sync endpoint. Backend explained that `clearHistory` is treated as custom data, and custom data is not sent on `/sync` call, so it's not sent. It's a bug that will be resolved by backend, but client needs a workaround meanwhile.

Also resolves that decoding errors are not printed to console.

* Fix unwatched channels are not hidden on new message

When the client starts for the first time, the channels linked to the query will be watched. If a channel was hidden on a previous session, that channel will not be linked to the query, so will not be watched. We won't receive a `MessageNew` event for that channel (since it's not being watched) and we won't mark it as `hidden = false`, so it won't show up.

The visibility middleware should also act on `Notification` events and mark the channel accordingly. Then it can be linked to any query that's currently alive.

* Fix hidden channel messages re-appearing

This happens because client uses `truncatedAt` as `hiddenAt`. Hiding channel with clearHistory flag sets truncatedAt, but backend is not aware so they send null for this value, and client uses null value to reset truncatedAt, effectively un-hiding the messages.

As per slack discussion (https://getstream.slack.com/archives/CE5N802GP/p1654094011610589?thread_ts=1654093897.531669&cid=CE5N802GP) it's not possible to un-truncate a channel by setting truncatedAt to null once it's set, so it's fine to only use non-nil truncatedAt values sent by backend

* Fix hidden channels' previewMessage is not updated

* Update CHANGELOG