Keep local reference to last fetched message in ChannelController to avoid broken pagination
When saving a ChannelPayload, it can contain Quoted and Pinned messages that might not be part of
the page that was fetched, but those are displayed as they meet all the current criterias of the
DB observer. But they can create gaps if we use them as the id for the next page.