[CIS-2264] Make sure to reload rows after moving them (#2371)
* Make sure to reload rows after moving them

UIKit does not reload rows when they are moved. This is really bad for us because when
we receive a .move update, it comes with changes to the cell itself. In most of the cases,
because of other events, this is not an issue. But in edge cases, the cell stays with the
old state until a new event comes in.

This PR ensures that whenever those .move events come, we always update the cell just in case.

* Update CHANGELOG